+++ src/scitacean/model.py
@@ -8,7 +8,19 @@
 
 """Pydantic models to encode data for communication with SciCat."""
 
-import enum
+try:
+    # Python 3.11+
+    from enum import StrEnum as _StrEnum
+
+    _DatasetTypeBases = (_StrEnum,)
+except ImportError:
+    from enum import Enum as _Enum
+
+    _DatasetTypeBases = (
+        str,
+        _Enum,
+    )
+
 import os
 from datetime import datetime
 from typing import Any, Dict, List, Optional, Type, TypeVar
@@ -21,8 +33,7 @@ from .logging import get_logger
 from .pid import PID
 
 
-# This can be replaced by StrEnum in Python 3.11+
-class DatasetType(str, enum.Enum):
+class DatasetType(*_DatasetTypeBases):
     """Type of Dataset"""
 
     RAW = "raw"