+++ src/scitacean/transfer/ssh.py
@@ -6,7 +6,7 @@ from contextlib import contextmanager
 from datetime import datetime, timedelta
 from getpass import getpass
 from pathlib import Path
-from typing import Any, Iterator, List, Optional, Tuple, Union
+from typing import Any, Callable, Iterator, List, Optional, Tuple, Union
 
 from dateutil.tz import tzoffset
 
@@ -23,7 +23,6 @@ from ..filesystem import RemotePath
 from ..logging import get_logger
 from .util import source_folder_for
 
-# TODO process multiple files together
 # TODO pass pid in put/revert?
 #      downloading does not need a pid, so it should not be required in the constructor/
 
@@ -44,7 +43,7 @@ class SSHDownloadConnection:
             self._connection.host,
             local,
         )
-        self._connection.get(remote=str(remote), local=str(local))
+        self._connection.get(remote=remote.posix, local=os.fspath(local))
 
 
 class SSHUploadConnection:
@@ -66,7 +65,9 @@ class SSHUploadConnection:
 
     def _make_source_folder(self) -> None:
         try:
-            self._connection.run(f"mkdir -p {os.fspath(self.source_folder)}", hide=True)
+            self._connection.run(
+                f"mkdir -p {self.source_folder.posix}", hide=True, in_stream=False
+            )
         except OSError as exc:
             raise FileUploadError(
                 f"Failed to create source folder {self.source_folder}: {exc.args}"
@@ -78,13 +79,16 @@ class SSHUploadConnection:
         uploaded = []
         try:
             for file in files:
-                uploaded.append(self._upload_file(file))
+                up, exc = self._upload_file(file)
+                uploaded.append(up)  # need to add this file in order to revert it
+                if exc is not None:
+                    raise exc
         except Exception:
             self.revert_upload(*uploaded)
             raise
         return uploaded
 
-    def _upload_file(self, file: File) -> File:
+    def _upload_file(self, file: File) -> Tuple[File, Optional[Exception]]:
         remote_path = self.remote_path(file.remote_path)
         get_logger().info(
             "Uploading file %s to %s on host %s",
@@ -92,29 +96,32 @@ class SSHUploadConnection:
             remote_path,
             self._connection.host,
         )
-        self._sftp.put(
-            remotepath=os.fspath(remote_path), localpath=os.fspath(file.local_path)
+        st = self._sftp.put(
+            remotepath=remote_path.posix, localpath=os.fspath(file.local_path)
         )
-        st = self._sftp.stat(os.fspath(remote_path))
-        self._validate_upload(file)
+        if (exc := self._validate_upload(file)) is not None:
+            return file, exc
         creation_time = (
             datetime.fromtimestamp(st.st_mtime, tz=self._remote_timezone)
             if st.st_mtime
             else None
         )
-        return file.uploaded(
-            remote_gid=str(st.st_gid),
-            remote_uid=str(st.st_uid),
-            remote_creation_time=creation_time,
-            remote_perm=str(st.st_mode),
-            remote_size=st.st_size,
+        return (
+            file.uploaded(
+                remote_gid=str(st.st_gid),
+                remote_uid=str(st.st_uid),
+                remote_creation_time=creation_time,
+                remote_perm=str(st.st_mode),
+                remote_size=st.st_size,
+            ),
+            None,
         )
 
-    def _validate_upload(self, file: File) -> None:
+    def _validate_upload(self, file: File) -> Optional[Exception]:
         if (checksum := self._compute_checksum(file)) is None:
-            return
+            return None
         if checksum != file.checksum():
-            raise FileUploadError(
+            return FileUploadError(
                 f"Upload of file {file.remote_path} failed: "
                 f"Checksum of uploaded file ({checksum}) does not "
                 f"match checksum of local file ({file.checksum()}) "
@@ -126,7 +133,9 @@ class SSHUploadConnection:
             return None
         try:
             res = self._connection.run(
-                f"{hash_exe} {self.remote_path(file.remote_path)}", hide=True
+                f"{hash_exe} {self.remote_path(file.remote_path).posix}",
+                hide=True,
+                in_stream=False,
             )
         except UnexpectedExit as exc:
             if exc.result.return_code == 127:
@@ -143,7 +152,9 @@ class SSHUploadConnection:
     def _get_remote_timezone(self) -> tzoffset:
         cmd = 'date +"%Z|%::z"'
         try:
-            tz_str = self._connection.run(cmd, hide=True).stdout.strip()
+            tz_str = self._connection.run(
+                cmd, hide=True, in_stream=False
+            ).stdout.strip()
         except UnexpectedExit as exc:
             raise FileUploadError(
                 f"Failed to get timezone of fileserver: {exc.args}"
@@ -166,7 +177,7 @@ class SSHUploadConnection:
                     self.source_folder,
                     self._connection.host,
                 )
-                self._sftp.rmdir(os.fspath(self.source_folder))
+                self._sftp.rmdir(self.source_folder.posix)
             except UnexpectedExit as exc:
                 get_logger().warning(
                     "Failed to remove empty remote directory %s on host:\n%s",
@@ -185,7 +196,7 @@ class SSHUploadConnection:
         )
 
         try:
-            self._sftp.remove(os.fspath(remote_path))
+            self._sftp.remove(remote_path.posix)
         except UnexpectedExit as exc:
             get_logger().warning(
                 "Error reverting file %s:\n%s", remote_path, exc.result
@@ -285,7 +296,9 @@ class SSHFileTransfer:
         self._host = host
         self._port = port
         self._source_folder_pattern = (
-            RemotePath(source_folder) if source_folder is not None else None
+            RemotePath(source_folder)
+            if isinstance(source_folder, str)
+            else source_folder
         )
 
     def source_folder_for(self, dataset: Dataset) -> RemotePath:
@@ -293,19 +306,56 @@ class SSHFileTransfer:
         return source_folder_for(dataset, self._source_folder_pattern)
 
     @contextmanager
-    def connect_for_download(self) -> Iterator[SSHDownloadConnection]:
-        """Create a connection for downloads, use as a context manager."""
-        con = _connect(self._host, self._port)
+    def connect_for_download(
+        self, connect: Optional[Callable[..., Connection]] = None
+    ) -> Iterator[SSHDownloadConnection]:
+        """Create a connection for downloads, use as a context manager.
+
+        Parameters
+        ----------
+        connect:
+            A function that creates and returns a :class:`fabric.connection.Connection`
+            object.
+            Will first be called with only ``host`` and ``port``.
+            If this fails (by raising
+            :class:`paramiko.ssh_exception.AuthenticationException`), the function is
+            called with ``host``, ``port``, and, optionally, ``user`` and
+            ``connection_kwargs`` depending on the authentication method.
+            Raising :class:`paramiko.ssh_exception.AuthenticationException` in the 2nd
+            call or any other exception in the 1st signals failure of
+            ``connect_for_download``.
+        """
+        con = _connect(self._host, self._port, connect=connect)
         try:
             yield SSHDownloadConnection(connection=con)
         finally:
             con.close()
 
     @contextmanager
-    def connect_for_upload(self, dataset: Dataset) -> Iterator[SSHUploadConnection]:
-        """Create a connection for uploads, use as a context manager."""
+    def connect_for_upload(
+        self, dataset: Dataset, connect: Optional[Callable[..., Connection]] = None
+    ) -> Iterator[SSHUploadConnection]:
+        """Create a connection for uploads, use as a context manager.
+
+        Parameters
+        ----------
+        dataset:
+            The connection will be used to upload files of this dataset.
+            Used to determine the target folder.
+        connect:
+            A function that creates and returns a :class:`fabric.connection.Connection`
+            object.
+            Will first be called with only ``host`` and ``port``.
+            If this fails (by raising
+            :class:`paramiko.ssh_exception.AuthenticationException`), the function is
+            called with ``host``, ``port``, and, optionally, ``user`` and
+            ``connection_kwargs`` depending on the authentication method.
+            Raising :class:`paramiko.ssh_exception.AuthenticationException` in the 2nd
+            call or any other exception in the 1st signals failure of
+            ``connect_for_upload``.
+        """
         source_folder = self.source_folder_for(dataset)
-        con = _connect(self._host, self._port)
+        con = _connect(self._host, self._port, connect=connect)
         try:
             yield SSHUploadConnection(
                 connection=con,
@@ -326,18 +376,31 @@ def _ask_for_credentials(host: str) -> Tuple[str, str]:
     return username, password
 
 
-def _generic_connect(host: str, port: Optional[int], **kwargs: Any) -> Connection:
-    con = Connection(host=host, port=port, **kwargs)
+def _generic_connect(
+    host: str,
+    port: Optional[int],
+    connect: Optional[Callable[..., Connection]],
+    **kwargs: Any,
+) -> Connection:
+    if connect is None:
+        con = Connection(host=host, port=port, **kwargs)
+    else:
+        con = connect(host=host, port=port, **kwargs)
     con.open()
     return con
 
 
-def _unauthenticated_connect(host: str, port: Optional[int]) -> Connection:
-    return _generic_connect(host=host, port=port)
+def _unauthenticated_connect(
+    host: str, port: Optional[int], connect: Optional[Callable[..., Connection]]
+) -> Connection:
+    return _generic_connect(host=host, port=port, connect=connect)
 
 
 def _authenticated_connect(
-    host: str, port: Optional[int], exc: AuthenticationException
+    host: str,
+    port: Optional[int],
+    connect: Optional[Callable[..., Connection]],
+    exc: AuthenticationException,
 ) -> Connection:
     # TODO fail fast if output going to file
     if isinstance(exc, PasswordRequiredException) and "encrypted" in exc.args[0]:
@@ -345,23 +408,30 @@ def _authenticated_connect(
         return _generic_connect(
             host=host,
             port=port,
+            connect=connect,
             connect_kwargs={"passphrase": _ask_for_key_passphrase()},
         )
     else:
         username, password = _ask_for_credentials(host)
         return _generic_connect(
-            host=host, port=port, user=username, connect_kwargs={"password": password}
+            host=host,
+            port=port,
+            connect=connect,
+            user=username,
+            connect_kwargs={"password": password},
         )
 
 
-def _connect(host: str, port: Optional[int]) -> Connection:
+def _connect(
+    host: str, port: Optional[int], connect: Optional[Callable[..., Connection]]
+) -> Connection:
     try:
         try:
-            return _unauthenticated_connect(host, port)
+            return _unauthenticated_connect(host, port, connect)
         except AuthenticationException as exc:
-            return _authenticated_connect(host, port, exc)
+            return _authenticated_connect(host, port, connect, exc)
     except Exception as exc:
-        # We pass secrets as arguments to functions called in this block and those
+        # We pass secrets as arguments to functions called in this block, and those
         # can be leaked through exception handlers. So catch all exceptions
         # and strip the backtrace up to this point to hide those secrets.
         raise type(exc)(exc.args) from None
@@ -371,7 +441,7 @@ def _connect(host: str, port: Optional[int]) -> Connection:
 
 def _folder_is_empty(con: Connection, path: RemotePath) -> bool:
     try:
-        ls: str = con.run(f"ls {path}", hide=True).stdout
+        ls: str = con.run(f"ls {path.posix}", hide=True, in_stream=False).stdout
         return ls == ""
     except UnexpectedExit:
         return False  # no further processing is needed in this case